Add ankiconnect import, all known words stuff now on server

+++ Jiten.Core/Data/User/UserKnownWord.cs
@@ -12,7 +12,7 @@ public class UserKnownWord
 {
     public string UserId { get; set; } = string.Empty;
     public int WordId { get; set; }
-    public int ReadingIndex { get; set; }
+    public byte ReadingIndex { get; set; }
     public DateTime LearnedDate { get; set; } = DateTime.UtcNow;
     public KnownState KnownState { get; set; } = KnownState.Known;
 }
